requirements.txt
requests>=2.31
python-dotenv>=1.0
google-ads>=24.0
